requirements.txt
fastapi
uvicorn
tensorflow==2.12.0
joblib
pillow
numpy
scikit-learn
